# apps/api/src/routes/me.ts
import { Hono } from 'hono';
import { and, eq, inArray, isNull } from 'drizzle-orm';
import {
  users as usersTable,
  sessions as sessionsTable,
  emailTokens,
  oauthAccounts,
  jobs as jobsTable,
  auditLog,
  categories as categoriesTable,
  expenses as expensesTable,
  importBatches as importBatchesTable,
  connectedAccounts as connectedAccountsTable,
  notionConnections,
  type Db,
} from '@desk/db';
import {
  MeResponse,
  PatchMeRequest,
  ChangeEmailRequest,
  ConfirmEmailRequest,
  DeleteMeRequest,
  type MeResponseT,
  type PatchMeResponseT,
  type SessionsResponseT,
  type ExportDocumentT,
  type ConfirmEmailResponseT,
} from '@desk/contracts';
import type { AppVariables, JobsDep } from '../app.js';
import type { SessionUser } from '../middleware/session.js';
import type { PasswordHasher } from '../adapters/password.js';
import type { Mailer } from '../adapters/mailer.js';
import type { SessionStore } from '../adapters/session-store.js';
import type { RateLimiter } from '../adapters/rate-limiter.js';
import { requireAuth } from '../lib/require-auth.js';
import { ApiError } from '../lib/api-error.js';
import { emailChangeMail } from '../mail/email-change.js';
import type { ConnectionsService } from '../services/connections.js';
import { createWidgetsService } from '../services/widgets.js';

export type MeRoutesDeps = {
  db: Db;
  hasher: PasswordHasher;
  mailer: Mailer;
  limiter: RateLimiter;
  sessionStore: SessionStore;
  clock: { now(): Date };
  appOrigin: string;
  /** Revokes every connected account at its provider before the DELETE /me cascade. */
  connections: ConnectionsService;
  jobs?: JobsDep;
};

function toUserResponse(u: SessionUser): MeResponseT['user'] {
  return {
    id: u.id,
    email: u.email,
    defaultCurrency: u.defaultCurrency,
    theme: u.theme,
    timeZone: u.timeZone,
    temperatureUnit: u.temperatureUnit === 'F' ? 'F' : 'C',
    onboardingCompletedAt: u.onboardingCompletedAt ? u.onboardingCompletedAt.toISOString() : null,
    emailVerified: u.emailVerifiedAt !== null,
    createdAt: u.createdAt.toISOString(),
  };
}

async function sha256Hex(input: string): Promise<string> {
  const digest = await crypto.subtle.digest('SHA-256', new TextEncoder().encode(input));
  return [...new Uint8Array(digest)].map((b) => b.toString(16).padStart(2, '0')).join('');
}

const EMAIL_CHANGE_PREFIX = 'email_change:';

/** GET/PATCH /me, /me/sessions, /me/export, DELETE /me, /me/email(/confirm), password/oauth removal. */
export function createMeRoutes(deps: MeRoutesDeps) {
  const { db, hasher, mailer, sessionStore, clock, appOrigin, limiter, connections, jobs } = deps;
  const app = new Hono<{ Variables: AppVariables }>();
  const widgetsService = createWidgetsService(db);

  app.get('/me', (c) => {
    const user = requireAuth(c);
    const body: MeResponseT = MeResponse.parse({ user: toUserResponse(user) });
    return c.json(body);
  });

  app.patch('/me', async (c) => {
    const user = requireAuth(c);
    const patchBody = PatchMeRequest.parse(await c.req.json());

    const patch: Partial<typeof usersTable.$inferInsert> = {};
    if (patchBody.theme !== undefined) patch.theme = patchBody.theme;
    if (patchBody.timeZone !== undefined) patch.timeZone = patchBody.timeZone;
    if (patchBody.temperatureUnit !== undefined) patch.temperatureUnit = patchBody.temperatureUnit;
    if (patchBody.onboardingCompletedAt !== undefined) {
      patch.onboardingCompletedAt = patchBody.onboardingCompletedAt
        ? new Date(patchBody.onboardingCompletedAt)
        : null;
    }

    let jobId: string | undefined;
    if (patchBody.defaultCurrency && patchBody.defaultCurrency !== user.defaultCurrency) {
      patch.defaultCurrency = patchBody.defaultCurrency;

      // Dedupe: a currency.change already queued/running for this user wins — the UI shows one
      // progress indicator, so a second request surfaces the running job's id instead of racing it.
      const [existing] = await db
        .select()
        .from(jobsTable)
        .where(
          and(
            eq(jobsTable.userId, user.id),
            eq(jobsTable.name, 'currency.change'),
            inArray(jobsTable.status, ['queued', 'running']),
          ),
        )
        .limit(1);

      if (existing) {
        jobId = existing.id;
      } else {
        const [inserted] = await db
          .insert(jobsTable)
          .values({
            name: 'currency.change',
            userId: user.id,
            payload: {
              userId: user.id,
              fromCurrency: user.defaultCurrency,
              toCurrency: patchBody.defaultCurrency,
              changeDate: clock.now().toISOString().slice(0, 10),
            },
          })
          .returning();
        jobId = inserted?.id;
      }
    }

    let updated = user;
    if (Object.keys(patch).length > 0) {
      const [row] = await db
        .update(usersTable)
        .set(patch)
        .where(eq(usersTable.id, user.id))
        .returning();
      if (row) updated = row;
    }

    const body: PatchMeResponseT = {
      user: toUserResponse(updated),
      ...(jobId ? { job: { id: jobId } } : {}),
    };
    return c.json(body);
  });

  app.get('/me/sessions', async (c) => {
    const user = requireAuth(c);
    const currentSession = c.get('session');
    const now = clock.now();

    const rows = await db
      .select()
      .from(sessionsTable)
      .where(and(eq(sessionsTable.userId, user.id), isNull(sessionsTable.revokedAt)));

    const body: SessionsResponseT = {
      sessions: rows
        .filter((r) => r.expiresAt > now)
        .map((r) => ({
          id: r.id,
          current: currentSession?.id === r.id,
          lastSeenAt: r.lastSeenAt.toISOString(),
          userAgent: r.userAgent,
        })),
    };
    return c.json(body);
  });

  app.delete('/me/sessions/:id', async (c) => {
    const user = requireAuth(c);
    const id = c.req.param('id');
    const [row] = await db.select().from(sessionsTable).where(eq(sessionsTable.id, id)).limit(1);
    if (!row || row.userId !== user.id) throw new ApiError('not_found', 'Session not found', 404);

    await db.update(sessionsTable).set({ revokedAt: new Date() }).where(eq(sessionsTable.id, id));
    return c.body(null, 204);
  });

  app.get('/me/export', async (c) => {
    const user = requireAuth(c);

    const [categories, expenses, importBatches, [notion], connectedAccountRows, widgets] =
      await Promise.all([
        db.select().from(categoriesTable).where(eq(categoriesTable.userId, user.id)),
        db.select().from(expensesTable).where(eq(expensesTable.userId, user.id)),
        db.select().from(importBatchesTable).where(eq(importBatchesTable.userId, user.id)),
        db.select().from(notionConnections).where(eq(notionConnections.userId, user.id)).limit(1),
        db
          .select({
            provider: connectedAccountsTable.provider,
            address: connectedAccountsTable.address,
            label: connectedAccountsTable.label,
            capabilities: connectedAccountsTable.capabilities,
            status: connectedAccountsTable.status,
            pausedAt: connectedAccountsTable.pausedAt,
          })
          .from(connectedAccountsTable)
          .where(eq(connectedAccountsTable.userId, user.id)),
        widgetsService.exportFor(user),
      ]);

    // ExportDocument (contracts) has no widgets field yet; the document is not re-parsed.
    const doc: ExportDocumentT & { widgets: typeof widgets } = {
      exportedAt: clock.now().toISOString(),
      user: toUserResponse(user),
      categories,
      expenses,
      importBatches,
      notion: notion
        ? { connected: true, direction: notion.direction, databaseId: notion.databaseId }
        : { connected: false, direction: null, databaseId: null },
      connections: connectedAccountRows.map((row) => ({
        provider: row.provider as 'google' | 'microsoft' | 'standards',
        address: row.address,
        label: row.label,
        capabilities: row.capabilities as ('mail' | 'calendar')[],
        status: row.pausedAt
          ? 'paused'
          : (row.status as 'connected' | 'reconnect_needed' | 'error'),
      })),
      widgets,
      version: 1,
    };

    const stream = new ReadableStream({
      start(controller) {
        controller.enqueue(new TextEncoder().encode(JSON.stringify(doc)));
        controller.close();
      },
    });
    return new Response(stream, { headers: { 'content-type': 'application/json' } });
  });

  app.delete('/me', async (c) => {
    const user = requireAuth(c);
    const raw = await c.req.json().catch(() => ({}));
    const body = DeleteMeRequest.parse(raw);

    if (user.passwordHash) {
      if (!body.password || !(await hasher.verify(body.password, user.passwordHash))) {
        throw new ApiError('validation_failed', 'Incorrect password', 400);
      }
    }

    // audit_log.user_id is ON DELETE SET NULL (data-model.md: audit rows must survive account
    // deletion) — pseudonymise with a one-way hash into `details` before the user row goes,
    // since the FK itself only nulls the column and knows nothing about pseudonymisation.
    const pseudonym = await sha256Hex(user.id);

    // contracts/api.md: DELETE /me revokes every connected account at its provider (best
    // effort, audited) before the cascade; a running/queued panels.refresh is cancelled first
    // so it can't race a revoked/soon-to-be-deleted account.
    await jobs?.cancelForUser?.(user.id);
    await connections.revokeAllForUser(user.id);

    await db.transaction(async (tx) => {
      await tx
        .update(jobsTable)
        .set({ status: 'failed', error: 'cancelled', finishedAt: new Date() })
        .where(
          and(eq(jobsTable.userId, user.id), inArray(jobsTable.status, ['queued', 'running'])),
        );

      const auditRows = await tx.select().from(auditLog).where(eq(auditLog.userId, user.id));
      for (const row of auditRows) {
        await tx
          .update(auditLog)
          .set({
            userId: null,
            details: {
              ...(row.details as Record<string, unknown> | null),
              deletedUserHash: pseudonym,
            },
          })
          .where(eq(auditLog.id, row.id));
      }

      await tx.delete(usersTable).where(eq(usersTable.id, user.id));
    });

    return c.body(null, 204);
  });

  app.post('/me/email', async (c) => {
    const user = requireAuth(c);
    const body = ChangeEmailRequest.parse(await c.req.json());

    const allowed = await limiter.hit(`email-change:${user.id}`, 5, 60 * 60 * 1000);
    if (!allowed) throw new ApiError('rate_limited', 'Too many attempts', 429);

    // Re-auth: a password-holding account MUST supply the correct password — it was previously
    // only checked when the caller bothered to send one, which is no check at all against
    // someone who found an already-signed-in session (e.g. a stolen cookie).
    if (user.passwordHash) {
      if (!body.password || !(await hasher.verify(body.password, user.passwordHash))) {
        throw new ApiError('validation_failed', 'Incorrect password', 400);
      }
    }

    const [existing] = await db
      .select({ id: usersTable.id })
      .from(usersTable)
      .where(eq(usersTable.email, body.newEmail))
      .limit(1);
    if (existing) throw new ApiError('conflict', 'Email already in use', 409);

    const token = crypto.randomUUID();
    const tokenHash = await sha256Hex(token);
    await db.insert(emailTokens).values({
      userId: user.id,
      purpose: `${EMAIL_CHANGE_PREFIX}${body.newEmail}`,
      tokenHash,
      expiresAt: new Date(clock.now().getTime() + 24 * 60 * 60 * 1000),
    });

    const confirmUrl = `${appOrigin}/me/email/confirm?token=${token}`;
    const template = emailChangeMail(confirmUrl);
    await mailer.send({ to: body.newEmail, subject: template.subject, html: template.html });
    await mailer.send({
      to: user.email,
      subject: 'Your email is changing',
      html: `<p>Your account email is changing to ${body.newEmail}. Contact support if this wasn't you.</p>`,
    });

    return c.body(null, 202);
  });

  app.post('/me/email/confirm', async (c) => {
    const user = requireAuth(c);
    const body = ConfirmEmailRequest.parse(await c.req.json());
    const tokenHash = await sha256Hex(body.token);

    const [row] = await db
      .select()
      .from(emailTokens)
      .where(eq(emailTokens.tokenHash, tokenHash))
      .limit(1);
    if (
      !row ||
      row.userId !== user.id ||
      row.usedAt ||
      row.expiresAt < clock.now() ||
      !row.purpose.startsWith(EMAIL_CHANGE_PREFIX)
    ) {
      throw new ApiError('validation_failed', 'Invalid or expired token', 400);
    }
    const newEmail = row.purpose.slice(EMAIL_CHANGE_PREFIX.length);

    let updated = user;
    await db.transaction(async (tx) => {
      await tx.update(emailTokens).set({ usedAt: new Date() }).where(eq(emailTokens.id, row.id));
      const [u] = await tx
        .update(usersTable)
        .set({ email: newEmail })
        .where(eq(usersTable.id, user.id))
        .returning();
      if (u) updated = u;
    });

    const currentSession = c.get('session');
    if (currentSession) await sessionStore.revokeAllExcept(user.id, currentSession.tokenHash);

    const respBody: ConfirmEmailResponseT = { user: toUserResponse(updated) };
    return c.json(respBody);
  });

  app.delete('/me/password', async (c) => {
    const user = requireAuth(c);
    const links = await db.select().from(oauthAccounts).where(eq(oauthAccounts.userId, user.id));
    if (links.length === 0) {
      throw new ApiError('conflict', 'Cannot remove your only sign-in method', 409);
    }
    await db.update(usersTable).set({ passwordHash: null }).where(eq(usersTable.id, user.id));
    return c.body(null, 204);
  });

  app.delete('/me/oauth/:provider', async (c) => {
    const user = requireAuth(c);
    const provider = c.req.param('provider');
    if (!user.passwordHash) {
      throw new ApiError('conflict', 'Cannot remove your only sign-in method', 409);
    }
    await db
      .delete(oauthAccounts)
      .where(and(eq(oauthAccounts.userId, user.id), eq(oauthAccounts.provider, provider)));
    return c.body(null, 204);
  });

  return app;
}