# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Desk API",
    "version": "0.0.0",
    "description": "Generated from packages/contracts/src via scripts/openapi.ts. Route list is hand-mapped from specs/001-phased-product-baseline/contracts/api.md; schemas are auto-converted from zod."
  },
  "paths": {
...
    "/me": {
      "get": {
        "summary": "Current user",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update profile/settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update profile/settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchMeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMeRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "See specs/001-phased-product-baseline/contracts/api.md"
          }
        }
      }
    },
    "/me/email": {
      "post": {
        "summary": "Request email change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "See specs/001-phased-product-baseline/contracts/api.md"
          }
        }
      }
    },
    "/me/email/confirm": {
      "post": {
        "summary": "Confirm email change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirm email change",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmEmailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeEmailRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "newEmail": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["newEmail"],
        "additionalProperties": false
      },
...
      "ConfirmEmailRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": ["token"],
        "additionalProperties": false
      },
      "ConfirmEmailResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "defaultCurrency": {
                "type": "string"
              },
              "theme": {
                "type": "string"
              },
              "timeZone": {
                "type": "string"
              },
              "onboardingCompletedAt": {
                "type": ["string", "null"]
              },
              "emailVerified": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string"
              },
              "temperatureUnit": {
                "type": "string",
                "enum": ["C", "F"]
              }
            },
            "required": [
              "id",
              "email",
              "defaultCurrency",
              "theme",
              "timeZone",
              "onboardingCompletedAt",
              "emailVerified",
              "createdAt",
              "temperatureUnit"
            ],
            "additionalProperties": false
          }
        },
        "required": ["user"],
        "additionalProperties": false
      },
...
      "DeleteMeRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "MeResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "defaultCurrency": {
                "type": "string"
              },
              "theme": {
                "type": "string"
              },
              "timeZone": {
                "type": "string"
              },
              "onboardingCompletedAt": {
                "type": ["string", "null"]
              },
              "emailVerified": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string"
              },
              "temperatureUnit": {
                "type": "string",
                "enum": ["C", "F"]
              }
            },
            "required": [
              "id",
              "email",
              "defaultCurrency",
              "theme",
              "timeZone",
              "onboardingCompletedAt",
              "emailVerified",
              "createdAt",
              "temperatureUnit"
            ],
            "additionalProperties": false
          }
        },
        "required": ["user"],
        "additionalProperties": false
      },
...
      "PatchMeRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "theme": {
            "type": "string"
          },
          "defaultCurrency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "timeZone": {
            "type": "string"
          },
          "onboardingCompletedAt": {
            "type": ["string", "null"]
          },
          "temperatureUnit": {
            "type": "string",
            "enum": ["C", "F"]
          }
        },
        "additionalProperties": false
      },
      "PatchMeResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "defaultCurrency": {
                "type": "string"
              },
              "theme": {
                "type": "string"
              },
              "timeZone": {
                "type": "string"
              },
              "onboardingCompletedAt": {
                "type": ["string", "null"]
              },
              "emailVerified": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string"
              },
              "temperatureUnit": {
                "type": "string",
                "enum": ["C", "F"]
              }
            },
            "required": [
              "id",
              "email",
              "defaultCurrency",
              "theme",
              "timeZone",
              "onboardingCompletedAt",
              "emailVerified",
              "createdAt",
              "temperatureUnit"
            ],
            "additionalProperties": false
          },
          "job": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": ["id"],
            "additionalProperties": false
          }
        },
        "required": ["user"],
        "additionalProperties": false
      },
...
    }
  }
}